Implemented Road Parent Class and Child Classes
Probably going to have to change around some of the formatting and logic because we might discover that it's better to have certain functions outside of the roads themselves.
For now though, roads have the ability to rotated and they contain the logic for checking if a placement is legal or not. Managing this would have to be done in the game manager currently though so thats why I think this might not be the best approach